pyproject.toml
[project]
name = "my-agent-practice"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "agent-framework",
    "python-dotenv",
    "mkdocs",
    "mkdocs-material",
    "fastapi>=0.133.0",
    "uvicorn>=0.41.0",
]

[tool.uv]
dev-dependencies = []

[tool.pylance]
# 可选：增强类型检查
typeCheckingMode = "standard"   # 或 "strict"
